Update Files
*Update MinCharge to version 4, based on partner's code
* Update TourFn to version 4, based on Partner's code

Known Bugs:
Round off Errors when a large number of Voronoi are in a close area.

Binary mask of a polygon has errors when dealing with borders.
Resolve: Use smaller steps.

MinCharge and Drone Path algorithm are not in sync. The distance drone travels are not taken into full account in MinCharge.