Reconstruction
Run.py: Change the script to OOP style. This form, the script is modular and can be used depending on the need.

MinCharge.py: Allow the user to specify if they want to use their own candidate CS instead of the generated. Also, instead of solving for 10 solutions in the linear program, we instead decrease the CS radius in order find a solution that  decrease the far away coverage in that field.

Field.py: The OOP style is not necessary since the methods are independent from each other and the structure does not bring any benefit to the overall project.

TestBench.py: Update the testbench to the new updates in the overall program.